Fix acpi_proc_read() not terminating properly
This caused a problem with e.g. `cat` producing garbled output
like: 0x2ealled

Here's why:
acpi_proc_read() is usually called multiple times in a row.
*off is supposed to be used by us to keep track of where we
are across those calls.
Once the user triggers a new read, *off will be 0 again.

Theoretically, we would need to be prepared to continue
reading from somewhere within the buffer on a second or
third call. But since ACPI calls don't tend to return a ton of data
and the target buffer appears to usually be kilobytes in size,
I think it is fair to gloss over this for now.